requirements.txt
torch
torchvision
datasets
numpy
pandas
opencv-python
scikit-image
matplotlib
tqdm
pillow
transformers
requests
accelerate
